PS C:\Users\ashis> cd github_projects/sokabe.github.io PS C:\Users\ashis\github_projects\sokabe.github.io> nano _config.yml nano : The term 'nano' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again. At line:1 char:1 + nano _config.yml + ~~~~     + CategoryInfo          : ObjectNotFound: (nano:String) [], CommandNotFoundException     + FullyQualifiedErrorId : CommandNotFoundException
PS C:\Users\ashis\github_projects\sokabe.github.io> vi _config.yml
vi : The term 'vi' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the
spelling of the name, or if a path was included, verify that the path is correct and try again.
At line:1 char:1
+ vi _config.yml
+ ~~
    + CategoryInfo          : ObjectNotFound: (vi:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException

PS C:\Users\ashis\github_projects\sokabe.github.io> bash -c “vi _config.yml”
bash : The term 'bash' is not recognized as the name of a cmdlet, function, script file, or operable program. Check
the spelling of the name, or if a path was included, verify that the path is correct and try again.
At line:1 char:1
+ bash -c “vi _config.yml”
+ ~~~~
    + CategoryInfo          : ObjectNotFound: (bash:String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException